+++ src/App.module.css
@@ -11,3 +11,11 @@
 .downloadButton {
     margin-top: 20px;
 }
+
+.buttonRow {
+    margin-top: 20px;
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+    gap: 10px;
+}
